Refactory: switching up to 32 classes (class_mask is uint32_t)

+++ user/C++/pfq.hpp
@@ -450,7 +450,7 @@ namespace net {
     // class mask
     //
     
-    typedef unsigned long class_mask;
+    typedef unsigned int class_mask;
 
     namespace 
     {
